# server/src/main/resources/static/swagger/openapi3.yaml
openapi: 3.0.1
info:
  title: Bannabe API
  description: Bannabe API 문서
  version: 0.0.1
servers:
- url: http://localhost:8080
...
paths:
...
  /v1/users/me:
    get:
      tags:
      - UserController
      summary: 회원 정보 조회(ME) API
      description: 회원 정보 조회(ME) API
      operationId: user_controller_test/get_user_info
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/get_user_info:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : {\r\n    \"email\" : \"test@test.com\"\
                    ,\r\n    \"nickname\" : \"TestUser\",\r\n    \"profileImage\"\
                    \ : \"https://test.com/test.jpg\",\r\n    \"isDefaultProfileImage\"\
                    \ : false\r\n  }\r\n}"
      security:
      - bearerAuthJWT: []
...
  /v1/users/me/nickname:
    patch:
      tags:
      - UserController
      summary: 닉네임 변경 API
      description: 닉네임 변경 API
      operationId: user_controller_test/change_nickname
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1-stations486549215"
            examples:
              user_controller_test/change_nickname:
                value: "{\r\n  \"nickname\" : \"newNickname\"\r\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/change_nickname:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : { }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/password:
    put:
      tags:
      - UserController
      summary: 비밀번호 재설정(로그인) API
      description: 비밀번호 재설정(로그인) API
      operationId: user_controller_test/change_password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1-stations486549215"
            examples:
              user_controller_test/change_password:
                value: "{\r\n  \"currentPassword\" : \"password\",\r\n  \"newPassword\"\
                  \ : \"newPassword\",\r\n  \"newPasswordConfirm\" : \"newPassword\"\
                  \r\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/change_password:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : { }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/profile-image:
    patch:
      tags:
      - UserController
      summary: 프로필 이미지 변경 API
      description: 프로필 이미지 변경 API
      operationId: user_controller_test/change_profile_image
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1-stations486549215"
            examples:
              user_controller_test/change_profile_image:
                value: "{\r\n  \"imageUrl\" : \"newProfileImage.png\"\r\n}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/change_profile_image:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : { }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/rentals:
    get:
      tags:
      - UserController
      summary: 대여 내역 조회 API
      description: 대여 내역 조회 API
      operationId: user_controller_test/get_all_rental_history
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/get_all_rental_history:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : {\r\n    \"content\" : [ {\r\n \
                    \     \"name\" : \"65W충전기\",\r\n      \"status\" : \"대여중\",\r\n\
                    \      \"rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-16T21:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-16T22:30:32\",\r\
                    \n      \"token\" : \"RH_hwZSgAlgzL\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"대여중\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-16T21:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-16T22:30:32\",\r\
                    \n      \"token\" : \"RH_ZGAJd6gyF1\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-15T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-15T21:30:32\",\r\
                    \n      \"token\" : \"RH_u88HJSo20p\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-14T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-14T21:30:32\",\r\
                    \n      \"token\" : \"RH_PVh1V1mrqB\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-13T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-13T21:30:32\",\r\
                    \n      \"token\" : \"RH_rZRYidxNsZ\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-12T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-12T21:30:32\",\r\
                    \n      \"token\" : \"RH_liS8I5fp5K\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-11T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-11T21:30:32\",\r\
                    \n      \"token\" : \"RH_k9sB5eut41\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-10T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-10T21:30:32\",\r\
                    \n      \"token\" : \"RH_bghqBVCWZq\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-09T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-09T21:30:32\",\r\
                    \n      \"token\" : \"RH_xvWNhqa07i\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"반납\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-08T20:30:32\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-08T21:30:32\",\r\
                    \n      \"token\" : \"RH_TOTZay3X3Q\"\r\n    } ],\r\n    \"page\"\
                    \ : {\r\n      \"size\" : 10,\r\n      \"number\" : 0,\r\n   \
                    \   \"totalElements\" : 20,\r\n      \"totalPages\" : 2\r\n  \
                    \  }\r\n  }\r\n}"
      security:
      - bearerAuthJWT: []
...
  /v1/users/me/profile-image/default:
    patch:
      tags:
      - UserController
      summary: 프로필 이미지 초기화 API
      description: 프로필 이미지 초기화 API
      operationId: user_controller_test/change_profile_image_to_default
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/change_profile_image_to_default:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : { }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/profile-image/pre-signed:
    get:
      tags:
      - UserController
      summary: S3 PreSignedURL 조회 API
      description: S3 PreSignedURL 조회 API
      operationId: user_controller_test/get_pre_signed_url
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/get_pre_signed_url:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : {\r\n    \"preSignedUrl\" : \"https://test.com/test.jpg\"\
                    \r\n  }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/rentals/active:
    get:
      tags:
      - UserController
      summary: 대여 현황 조회 API
      description: 대여 현황 조회 API
      operationId: user_controller_test/get_active_rental_history
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/get_active_rental_history:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : {\r\n    \"rentals\" : [ {\r\n \
                    \     \"name\" : \"65W충전기\",\r\n      \"status\" : \"대여중\",\r\n\
                    \      \"rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-16T21:30:31\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-16T22:30:31\",\r\
                    \n      \"token\" : \"RH_AMeXSnIKca\"\r\n    }, {\r\n      \"\
                    name\" : \"65W충전기\",\r\n      \"status\" : \"대여중\",\r\n      \"\
                    rentalTimeHour\" : 1,\r\n      \"startTime\" : \"2025-05-16T21:30:31\"\
                    ,\r\n      \"expectedReturnTime\" : \"2025-05-16T22:30:31\",\r\
                    \n      \"token\" : \"RH_Rmy2ef9GUy\"\r\n    } ]\r\n  }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/stations/bookmark:
    get:
      tags:
      - UserController
      summary: 북마크 스테이션 조회 API
      description: 북마크 스테이션 조회 API
      operationId: user_controller_test/get_bookmark_stations
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/get_bookmark_stations:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : {\r\n    \"bookmarks\" : [ {\r\n\
                    \      \"name\" : \"테스트 스테이션0\",\r\n      \"stationId\" : 2,\r\
                    \n      \"bookmarkId\" : 2\r\n    }, {\r\n      \"name\" : \"테\
                    스트 스테이션1\",\r\n      \"stationId\" : 3,\r\n      \"bookmarkId\"\
                    \ : 3\r\n    }, {\r\n      \"name\" : \"테스트 스테이션2\",\r\n     \
                    \ \"stationId\" : 4,\r\n      \"bookmarkId\" : 4\r\n    }, {\r\
                    \n      \"name\" : \"테스트 스테이션3\",\r\n      \"stationId\" : 5,\r\
                    \n      \"bookmarkId\" : 5\r\n    }, {\r\n      \"name\" : \"테\
                    스트 스테이션4\",\r\n      \"stationId\" : 6,\r\n      \"bookmarkId\"\
                    \ : 6\r\n    }, {\r\n      \"name\" : \"테스트 스테이션5\",\r\n     \
                    \ \"stationId\" : 7,\r\n      \"bookmarkId\" : 7\r\n    }, {\r\
                    \n      \"name\" : \"테스트 스테이션6\",\r\n      \"stationId\" : 8,\r\
                    \n      \"bookmarkId\" : 8\r\n    }, {\r\n      \"name\" : \"테\
                    스트 스테이션7\",\r\n      \"stationId\" : 9,\r\n      \"bookmarkId\"\
                    \ : 9\r\n    }, {\r\n      \"name\" : \"테스트 스테이션8\",\r\n     \
                    \ \"stationId\" : 10,\r\n      \"bookmarkId\" : 10\r\n    }, {\r\
                    \n      \"name\" : \"테스트 스테이션9\",\r\n      \"stationId\" : 11,\r\
                    \n      \"bookmarkId\" : 11\r\n    } ]\r\n  }\r\n}"
      security:
      - bearerAuthJWT: []
  /v1/users/me/stations/bookmark/{bookmarkId}:
    delete:
      tags:
      - UserController
      summary: 북마크 스테이션 삭제 API
      description: 북마크 스테이션 삭제 API
      operationId: user_controller_test/remove_bookmark_station
      parameters:
      - name: bookmarkId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1-stations486549215"
              examples:
                user_controller_test/remove_bookmark_station:
                  value: "{\r\n  \"success\" : true,\r\n  \"message\" : \"요청이 성공적으\
                    로 처리되었습니다.\",\r\n  \"data\" : { }\r\n}"
      security:
      - bearerAuthJWT: []
components:
  schemas:
    v1-stations486549215:
      type: object
  securitySchemes:
    bearerAuthJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT